move price to shared folder
